feat: `&+` operator
Add supports for `+` for Array based sequencing and `&+` for sequencing
Parsers as a tuple. This provides better compatability with complex
parsing as the type system tops out at Tuples of size 300.

Additionally this is intended to provide a more uniform API for when
composing parsers so that repitition and sequencing both output results
as an Array.